requirements.txt
django>=1.11,<1.12
django-debug-toolbar>=1.4,<2.0
django-rest-swagger>=2.1,<2.2
djangorestframework>=3.6,<3.7
factory_boy>=2.8,<2.9
social-auth-app-django>=1.1,<1.2
